feat: invite user profile page rename to match standards
