+++ client/app/profile/components/Avatar.tsx
