+++ client/app/profile/components/FlatshareDetails.tsx
